COM-76: Added  support for lambda worker to Go image.
