+++ cmd/cli/prepare_worker.go
@@ -42,7 +42,6 @@ type workerBuilder struct {
 func (b *workerBuilder) addCLIFlags(fs *pflag.FlagSet) {
 	fs.StringVar(&b.DirName, "dir-name", "", "Directory name for prepared worker")
 	fs.StringVar(&b.ProjectName, "project-name", "", "Name of project (builds a project instead of standard worker)")
-	fs.BoolVar(&b.Lambda, "lambda", false, "Build a Lambda-compatible worker (reads config from environment variables, no CLI args)")
 	b.SdkOptions.AddCLIFlags(fs)
 	fs.AddFlagSet(b.loggingOptions.FlagSet())
 }
